Refactor task-specific modules to lib/tasks/
