+++ demo3.py
@@ -33,7 +33,7 @@ import random
 from ec import commandlineArguments, ecIterator
 from grammar import Grammar
 from program import Primitive
-from task import Task
+from lib.tasks.task import Task
 from type import arrow, tint
 from utilities import numberOfCPUs
