+++ enumeration.py
@@ -1,7 +1,7 @@
 from utilities import eprint
 from likelihoodModel import AllOrNothingLikelihoodModel
 from frontier import *
-from task import *
+from lib.tasks.task import *
 from type import *
 from program import *
 from grammar import *
