+++ lib/tasks/makePuddleworldTasks.py
@@ -4,8 +4,7 @@ Tasks are (gridworld, text instruction) -> goal coordinate.
 Credit: tasks are taken from: https://github.com/JannerM/spatial-reasoning 
 """
 from puddleworldPrimitives import *
-from utilities import *
-from task import *
+from lib.tasks.task import *
 from type import *
 
 OBJECT_NAMES = ["NULL", "puddle", "star", "circle", "triangle", "heart", "spade", "diamond", "rock", "tree", "house", "horse"]
